% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEPTED-package.R
\name{KEPTED}
\alias{KEPTED}
\title{Provides an implementation of a kernel-embedding of probability test for
elliptical distribution. This is an asymptotic test for elliptical
distribution under general alternatives, and the location and shape
parameters are assumed to be unknown. Some side-products are posted,
including the transformation between rectangular and polar coordinates and
two product-type kernel functions.}
\description{
Provides an implementation of a kernel-embedding of probability test for
elliptical distribution. This is an asymptotic test for elliptical
distribution under general alternatives, and the location and shape
parameters are assumed to be unknown. Some side-products are posted,
including the transformation between rectangular and polar coordinates and
two product-type kernel functions.
}
\references{
\cite{Tang, Y. and Li, B. (2024), “A nonparametric test for elliptical
distribution based on kernel embedding of probabilities,”
\url{https://arxiv.org/abs/2306.10594}}
}
\author{
Yin Tang \email{yqt5219@psu.edu}
Bing Li \email{bxl9@psu.edu}
}
